// src/api/generated/models/termsResDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Fittner API
 * Fittner API with Spring Boot
 * OpenAPI spec version: 1.0.0
 */

/**
 * 응답 데이터
 */
export interface TermsResDto {
  /** 약관필수여부 */
  termsEssentialYn?: string;
  /** 약관ID */
  termsId?: string;
  /** 약관종류 */
  termsKind?: string;
  /** 약관제목 */
  termsTitle?: string;
  /** 약관url */
  termsUrl?: string;
}